+++ CHANGELOG.md
@@ -13,7 +13,10 @@ File structure as changed and tinyUSB was modified and integrated to allow compi
 RC2 adds/fixes the following issue:
 
 ### Added
-  - added SAMD21 digital input ISR (#89)
+  - implemented SAMD21 digital input ISR (#89)
+  - implemented STM32F1 analog input reading (#88)
+  - implemented io_control analog reading function (#88)
+  - implemented SAMD21 analog reading function (#90)
 
 ### Changed
   - modified file structure and tinyUSB was modified and integrated to allow compiling code using Arduino IDE (#87)